Restructure public interface
